requirements.txt
psutil==5.9.8
numpy-stl
matplotlib
numpy
Pillow
